Crate cognitive_services_speech_sdk_rs

cognitive-services-speech-sdk-rs

Pure Rust binding for the Microsoft Cognitive Service Speech SDK. Provides thin abstraction around native C API. Heavily inspired by official Golang library.

For more information about Micorost Speech Service see here.

Modules

audio

Package audio provides the audio configuration, input/output streams, and related utilities for audio interactions.

common

Package common contains the definitions for many of the shared objects and properties in the Speech SDK.

dialog

Package dialog provides functionality for creating custom voice assistant applications and managing the related interaction flow.

error

Package error contains struct Error used to wrap library errors.

ffi

Package ffi contains bindings to underlying C API and thin abstraction for managing C handles.

speech

Package speech provides functionality for speech recognizers along with their related configuration and event objects.