Enum cognitive_services_speech_sdk_rs::common::PropertyId

pub enum PropertyId {

    SpeechServiceConnectionKey = 1_000,
    SpeechServiceConnectionEndpoint = 1_001,
    SpeechServiceConnectionRegion = 1_002,
    SpeechServiceAuthorizationToken = 1_003,
    SpeechServiceAuthorizationType = 1_004,
    SpeechServiceConnectionEndpointId = 1_005,
    SpeechServiceConnectionHost = 1_006,
    SpeechServiceConnectionProxyHostName = 1_100,
    SpeechServiceConnectionProxyPort = 1_101,
    SpeechServiceConnectionProxyUserName = 1_102,
    SpeechServiceConnectionProxyPassword = 1_103,
    SpeechServiceConnectionURL = 1_104,
    SpeechServiceConnectionTranslationToLanguages = 2_000,
    SpeechServiceConnectionTranslationVoice = 2_001,
    SpeechServiceConnectionTranslationFeatures = 2_002,
    SpeechServiceConnectionIntentRegion = 2_003,
    SpeechServiceConnectionRecoMode = 3_000,
    SpeechServiceConnectionRecoLanguage = 3_001,
    SpeechSessionId = 3_002,
    SpeechServiceConnectionUserDefinedQueryParameters = 3_003,
    SpeechServiceConnectionRecoModelName = 3_005,
    SpeechServiceConnectionRecoModelKey = 3_006,
    SpeechServiceConnectionSynthLanguage = 3_100,
    SpeechServiceConnectionSynthVoice = 3_101,
    SpeechServiceConnectionSynthOutputFormat = 3_102,
    SpeechServiceConnectionSynthEnableCompressedAudioTransmission = 3_103,
    SpeechServiceConnectionSynthOfflineVoice = 3_113,
    SpeechServiceConnectionSynthModelKey = 3_114,
    SpeechServiceConnectionInitialSilenceTimeoutMs = 3_200,
    SpeechServiceConnectionEndSilenceTimeoutMs = 3_201,
    SpeechServiceConnectionEnableAudioLogging = 3_202,
    SpeechServiceResponseRequestDetailedResultTrueFalse = 4_000,
    SpeechServiceResponseRequestProfanityFilterTrueFalse = 4_001,
    SpeechServiceResponseProfanityOption = 4_002,
    SpeechServiceResponsePostProcessingOption = 4_003,
    SpeechServiceResponseRequestWordLevelTimestamps = 4_004,
    SpeechServiceResponseStablePartialResultThreshold = 4_005,
    SpeechServiceResponseOutputFormatOption = 4_006,
    SpeechServiceResponseTranslationRequestStablePartialResult = 4_100,
    SpeechServiceResponseJsonResult = 5_000,
    SpeechServiceResponseJsonErrorDetails = 5_001,
    SpeechServiceResponseRecognitionLatencyMs = 5_002,
    SpeechServiceResponseSynthesisFirstByteLatencyMs = 5_010,
    SpeechServiceResponseSynthesisFinishLatencyMs = 5_011,
    SpeechServiceResponseSynthesisUnderrunTimeMs = 5_012,
    SpeechServiceResponseSynthesisBackend = 5_020,
    CancellationDetailsReason = 6_000,
    CancellationDetailsReasonText = 6_001,
    CancellationDetailsReasonDetailedText = 6_002,
    LanguageUnderstandingServiceResponseJsonResult = 7_000,
    AudioConfigDeviceNameForCapture = 8_000,
    AudioConfigNumberOfChannelsForCapture = 8_001,
    AudioConfigSampleRateForCapture = 8_002,
    AudioConfigBitsPerSampleForCapture = 8_003,
    AudioConfigAudioSource = 8_004,
    AudioConfigDeviceNameForRender = 8_005,
    AudioConfigPlaybackBufferLengthInMs = 8_006,
    SpeechLogFilename = 9_001,
    ConversationApplicationID = 10_000,
    ConversationDialogType = 10_001,
    ConversationInitialSilenceTimeout = 10_002,
    ConversationFromID = 10_003,
    ConversationConversationID = 10_004,
    ConversationCustomVoiceDeploymentIDs = 10_005,
    ConversationSpeechActivityTemplate = 10_006,
    DataBufferTimeStamp = 11_001,
    DataBufferUserID = 11_002,

}

PropertyID defines speech property ids.

Variants

SpeechServiceConnectionKey = 1_000

SpeechServiceConnectionKey is the Cognitive Services Speech Service subscription key. If you are using an intent recognizer, you need to specify the LUIS endpoint key for your particular LUIS app. Under normal circumstances, you shouldn’t have to use this property directly. Instead, use NewSpeechConfigFromSubscription.

SpeechServiceConnectionEndpoint = 1_001

SpeechServiceConnectionEndpoint is the Cognitive Services Speech Service endpoint (url). Under normal circumstances, you shouldn’t have to use this property directly. Instead, use NewSpeechConfigFromEndpoint. NOTE: This endpoint is not the same as the endpoint used to obtain an access token.

SpeechServiceConnectionRegion = 1_002

SpeechServiceConnectionRegion is the Cognitive Services Speech Service region. Under normal circumstances, you shouldn’t have to use this property directly. Instead, use NewSpeechConfigFromSubscription, NewSpeechConfigFromEndpoint, NewSpeechConfigFromHost, NewSpeechConfigFromAuthorizationToken.

SpeechServiceAuthorizationToken = 1_003

SpeechServiceAuthorizationToken is the Cognitive Services Speech Service authorization token (aka access token). Under normal circumstances, you shouldn’t have to use this property directly. Instead, use NewSpeechConfigFromAuthorizationToken, Recognizer.SetAuthorizationToken

SpeechServiceAuthorizationType = 1_004

SpeechServiceConnectionEndpointId = 1_005

SpeechServiceConnectionEndpointID is the Cognitive Services Custom Speech Service endpoint id. Under normal circumstances, you shouldn’t have to use this property directly. Instead use SpeechConfig.SetEndpointId. NOTE: The endpoint id is available in the Custom Speech Portal, listed under Endpoint Details.

SpeechServiceConnectionHost = 1_006

SpeechServiceConnectionHost is the Cognitive Services Speech Service host (url). Under normal circumstances, you shouldn’t have to use this property directly. Instead, use NewSpeechConfigFromHost.

SpeechServiceConnectionProxyHostName = 1_100

SpeechServiceConnectionProxyHostName is the host name of the proxy server used to connect to the Cognitive Services Speech Service. Under normal circumstances, you shouldn’t have to use this property directly. Instead, use SpeechConfig.SetProxy.

SpeechServiceConnectionProxyPort = 1_101

SpeechServiceConnectionProxyPort is the port of the proxy server used to connect to the Cognitive Services Speech Service. Under normal circumstances, you shouldn’t have to use this property directly. Instead, use SpeechConfig.SetProxy.

SpeechServiceConnectionProxyUserName = 1_102

SpeechServiceConnectionProxyUserName is the user name of the proxy server used to connect to the Cognitive Services Speech Service. Under normal circumstances, you shouldn’t have to use this property directly. Instead, use SpeechConfig.SetProxy.

SpeechServiceConnectionProxyPassword = 1_103

SpeechServiceConnectionProxyPassword is the password of the proxy server used to connect to the Cognitive Services Speech Service. Under normal circumstances, you shouldn’t have to use this property directly. Instead, use SpeechConfig.SetProxy.

SpeechServiceConnectionURL = 1_104

SpeechServiceConnectionURL is the URL string built from speech configuration. This property is intended to be read-only. The SDK is using it internally.

SpeechServiceConnectionTranslationToLanguages = 2_000

SpeechServiceConnectionTranslationToLanguages is the list of comma separated languages used as target translation languages. Under normal circumstances, you shouldn’t have to use this property directly. Instead use SpeechTranslationConfig.AddTargetLanguage and SpeechTranslationConfig.GetTargetLanguages.

SpeechServiceConnectionTranslationVoice = 2_001

SpeechServiceConnectionTranslationVoice is the name of the Cognitive Service Text to Speech Service voice. Under normal circumstances, you shouldn’t have to use this property directly. Instead use SpeechTranslationConfig.SetVoiceName. NOTE: Valid voice names can be found at https:///aka.ms/csspeech/voicenames.

SpeechServiceConnectionTranslationFeatures = 2_002

SpeechServiceConnectionTranslationFeatures is the translation features. For internal use.

SpeechServiceConnectionIntentRegion = 2_003

SpeechServiceConnectionIntentRegion is the Language Understanding Service region. Under normal circumstances, you shouldn’t have to use this property directly. Instead use LanguageUnderstandingModel.

SpeechServiceConnectionRecoMode = 3_000

SpeechServiceConnectionRecoMode is the Cognitive Services Speech Service recognition mode. Can be “INTERACTIVE”, “CONVERSATION” or “DICTATION”. This property is intended to be read-only. The SDK is using it internally.

SpeechServiceConnectionRecoLanguage = 3_001

SpeechServiceConnectionRecoLanguage is the spoken language to be recognized (in BCP-47 format). Under normal circumstances, you shouldn’t have to use this property directly. Instead, use SpeechConfig.SetSpeechRecognitionLanguage.

SpeechSessionId = 3_002

SpeechSessionID is the session id. This id is a universally unique identifier (aka UUID) representing a specific binding of an audio input stream and the underlying speech recognition instance to which it is bound. Under normal circumstances, you shouldn’t have to use this property directly.

SpeechServiceConnectionUserDefinedQueryParameters = 3_003

SpeechServiceConnectionUserDefinedQueryParameters are the query parameters provided by users. They will be passed to the service as URL query parameters.

SpeechServiceConnectionRecoModelName = 3_005

The name of the model to be used for speech recognition. Under normal circumstances, you shouldn’t use this property directly. Currently this is only valid when EmbeddedSpeechConfig is used. Added in version 1.19.0

SpeechServiceConnectionRecoModelKey = 3_006

The decryption key of the model to be used for speech recognition. Under normal circumstances, you shouldn’t use this property directly. Currently this is only valid when EmbeddedSpeechConfig is used. Added in version 1.19.0

SpeechServiceConnectionSynthLanguage = 3_100

SpeechServiceConnectionSynthVoice = 3_101

SpeechServiceConnectionSynthVoice is the name of the TTS voice to be used for speech synthesis

SpeechServiceConnectionSynthOutputFormat = 3_102

SpeechServiceConnectionSynthOutputFormat is the string to specify TTS output audio format.

SpeechServiceConnectionSynthEnableCompressedAudioTransmission = 3_103

SpeechServiceConnectionSynthEnableCompressedAudioTransmission indicates if use compressed audio format for speech synthesis audio transmission. This property only affects when SpeechServiceConnectionSynthOutputFormat is set to a pcm format. If this property is not set and GStreamer is available, SDK will use compressed format for synthesized audio transmission, and decode it. You can set this property to “false” to use raw pcm format for transmission on wire.

SpeechServiceConnectionSynthOfflineVoice = 3_113

The name of the offline TTS voice to be used for speech synthesis Under normal circumstances, you shouldn’t use this property directly. Added in version 1.19.0

SpeechServiceConnectionSynthModelKey = 3_114

The decryption key of the voice to be used for speech synthesis. Under normal circumstances, you shouldn’t use this property directly. Added in version 1.19.0

SpeechServiceConnectionInitialSilenceTimeoutMs = 3_200

SpeechServiceConnectionInitialSilenceTimeoutMs is the initial silence timeout value (in milliseconds) used by the service.

SpeechServiceConnectionEndSilenceTimeoutMs = 3_201

SpeechServiceConnectionEndSilenceTimeoutMs is the end silence timeout value (in milliseconds) used by the service.

SpeechServiceConnectionEnableAudioLogging = 3_202

SpeechServiceConnectionEnableAudioLogging is a boolean value specifying whether audio logging is enabled in the service or not.

SpeechServiceResponseRequestDetailedResultTrueFalse = 4_000

SpeechServiceResponseRequestDetailedResultTrueFalse the requested Cognitive Services Speech Service response output format (simple or detailed). Under normal circumstances, you shouldn’t have to use this property directly. Instead use SpeechConfig.SetOutputFormat.

SpeechServiceResponseRequestProfanityFilterTrueFalse = 4_001

SpeechServiceResponseRequestProfanityFilterTrueFalse is the requested Cognitive Services Speech Service response output profanity level. Currently unused.

SpeechServiceResponseProfanityOption = 4_002

SpeechServiceResponseProfanityOption is the requested Cognitive Services Speech Service response output profanity setting. Allowed values are “masked”, “removed”, and “raw”.

SpeechServiceResponsePostProcessingOption = 4_003

SpeechServiceResponsePostProcessingOption a string value specifying which post processing option should be used by the service. Allowed values are “TrueText”.

SpeechServiceResponseRequestWordLevelTimestamps = 4_004

SpeechServiceResponseRequestWordLevelTimestamps is a boolean value specifying whether to include word-level timestamps in the response result.

SpeechServiceResponseStablePartialResultThreshold = 4_005

SpeechServiceResponseStablePartialResultThreshold is the number of times a word has to be in partial results to be returned.

SpeechServiceResponseOutputFormatOption = 4_006

SpeechServiceResponseOutputFormatOption is a string value specifying the output format option in the response result. Internal use only.

SpeechServiceResponseTranslationRequestStablePartialResult = 4_100

SpeechServiceResponseTranslationRequestStablePartialResult is a boolean value to request for stabilizing translation partial results by omitting words in the end.

SpeechServiceResponseJsonResult = 5_000

SpeechServiceResponseJSONResult is the Cognitive Services Speech Service response output (in JSON format). This property is available on recognition result objects only.

SpeechServiceResponseJsonErrorDetails = 5_001

SpeechServiceResponseJSONErrorDetails is the Cognitive Services Speech Service error details (in JSON format). Under normal circumstances, you shouldn’t have to use this property directly. Instead, use CancellationDetails.ErrorDetails.

SpeechServiceResponseRecognitionLatencyMs = 5_002

SpeechServiceResponseRecognitionLatencyMs is the recognition latency in milliseconds. Read-only, available on final speech/translation/intent results. This measures the latency between when an audio input is received by the SDK, and the moment the final result is received from the service. The SDK computes the time difference between the last audio fragment from the audio input that is contributing to the final result, and the time the final result is received from the speech service.

SpeechServiceResponseSynthesisFirstByteLatencyMs = 5_010

SpeechServiceResponseSynthesisFirstByteLatencyMs is the speech synthesis first byte latency in milliseconds. Read-only, available on final speech synthesis results. This measures the latency between when the synthesis is started to be processed, and the moment the first byte audio is available. Added in version 1.17.0.

SpeechServiceResponseSynthesisFinishLatencyMs = 5_011

SpeechServiceResponseSynthesisFinishLatencyMs is the speech synthesis all bytes latency in milliseconds. Read-only, available on final speech synthesis results. This measures the latency between when the synthesis is started to be processed, and the moment the whole audio is synthesized. Added in version 1.17.0.

SpeechServiceResponseSynthesisUnderrunTimeMs = 5_012

SpeechServiceResponseSynthesisUnderrunTimeMs is the underrun time for speech synthesis in milliseconds. Read-only, available on results in SynthesisCompleted events. This measures the total underrun time from AudioConfigPlaybackBufferLengthInMs is filled to synthesis completed. Added in version 1.17.0.

SpeechServiceResponseSynthesisBackend = 5_020

SpeechServiceResponseSynthesisBackend indicates which backend the synthesis is finished by. Read-only, available on speech synthesis results, except for the result in SynthesisStarted event Added in version 1.17.0.

CancellationDetailsReason = 6_000

CancellationDetailsReason is the cancellation reason. Currently unused.

CancellationDetailsReasonText = 6_001

CancellationDetailsReasonText the cancellation text. Currently unused.

CancellationDetailsReasonDetailedText = 6_002

CancellationDetailsReasonDetailedText is the cancellation detailed text. Currently unused.

LanguageUnderstandingServiceResponseJsonResult = 7_000

LanguageUnderstandingServiceResponseJSONResult is the Language Understanding Service response output (in JSON format). Available via IntentRecognitionResult.Properties.

AudioConfigDeviceNameForCapture = 8_000

AudioConfigDeviceNameForCapture is the device name for audio capture. Under normal circumstances, you shouldn’t have to use this property directly. Instead, use AudioConfig.FromMicrophoneInput.

AudioConfigNumberOfChannelsForCapture = 8_001

AudioConfigNumberOfChannelsForCapture is the number of channels for audio capture. Internal use only.

AudioConfigSampleRateForCapture = 8_002

AudioConfigSampleRateForCapture is the sample rate (in Hz) for audio capture. Internal use only.

AudioConfigBitsPerSampleForCapture = 8_003

AudioConfigBitsPerSampleForCapture is the number of bits of each sample for audio capture. Internal use only.

AudioConfigAudioSource = 8_004

AudioConfigAudioSource is the audio source. Allowed values are “Microphones”, “File”, and “Stream”.

AudioConfigDeviceNameForRender = 8_005

you shouldn’t have to use this property directly. Instead, use NewAudioConfigFromDefaultSpeakerOutput. Added in version 1.17.0

AudioConfigPlaybackBufferLengthInMs = 8_006

AudioConfigPlaybackBufferLengthInMs indicates the playback buffer length in milliseconds, default is 50 milliseconds.

SpeechLogFilename = 9_001

SpeechLogFilename is the file name to write logs.

ConversationApplicationID = 10_000

ConversationApplicationID is the identifier used to connect to the backend service.

ConversationDialogType = 10_001

ConversationDialogType is the type of dialog backend to connect to.

ConversationInitialSilenceTimeout = 10_002

ConversationInitialSilenceTimeout is the silence timeout for listening.

ConversationFromID = 10_003

ConversationFromID is the FromId to be used on speech recognition activities.

ConversationConversationID = 10_004

ConversationConversationID is the ConversationId for the session.

ConversationCustomVoiceDeploymentIDs = 10_005

ConversationCustomVoiceDeploymentIDs is a comma separated list of custom voice deployment ids.

ConversationSpeechActivityTemplate = 10_006

ConversationSpeechActivityTemplate is use to stamp properties in the template on the activity generated by the service for speech.

DataBufferTimeStamp = 11_001

DataBufferTimeStamp is the time stamp associated to data buffer written by client when using Pull/Push audio input streams. The time stamp is a 64-bit value with a resolution of 90 kHz. It is the same as the presentation timestamp in an MPEG transport stream. See https:///en.wikipedia.org/wiki/Presentation_timestamp

DataBufferUserID = 11_002

DataBufferUserID is the user id associated to data buffer written by client when using Pull/Push audio input streams.

Implementations

impl PropertyId

pub fn to_i32(&self) -> i32