Module cognitive_services_speech_sdk_rs::dialog

Package dialog provides functionality for creating custom voice assistant applications and managing the related interaction flow.

Structs

ActivityReceivedEvent

Event passed into callback registered by DialogServiceConnector::set_activity_received_cb.

BotFrameworkConfig

BotFrameworkConfig defines configurations for the dialog service connector object for using a Bot Framework backend.

CustomCommandsConfig

CustomCommandsConfig defines configurations for the dialog service connector object for using a CustomCommands backend.

DialogServiceConnector

DialogServiceConnector connects to a speech enabled dialog backend.

Traits

DialogServiceConfig

DialogServiceConfig defines base configurations for the dialog service connector and custom commands config objects.