Struct cognitive_services_speech_sdk_rs::dialog::CustomCommandsConfig

pub struct CustomCommandsConfig {
    pub config: SpeechConfig,
}

CustomCommandsConfig defines configurations for the dialog service connector object for using a CustomCommands backend.

Fields
config: SpeechConfig
Implementations

impl CustomCommandsConfig

pub fn from_subscription<S>( application_id: S, subscription_key: S, region: S ) -> Result<CustomCommandsConfig>
where S: Into<Vec<u8>>,

Creates a Custom Commands config instance with the specified application id, subscription key and region.

pub fn from_auth_token<S>( application_id: S, authorization_token: S, region: S ) -> Result<CustomCommandsConfig>
where S: Into<Vec<u8>>,

Creates a Custom Commands config instance with the specified application id authorization token and region. Note: The caller needs to ensure that the authorization token is valid. Before the authorization token expires, the caller needs to refresh it by calling this setter with a new valid token. As configuration values are copied when creating a new connector, the new token value will not apply to connectors that have already been created. For connectors that have been created before, you need to set authorization token of the corresponding connector to refresh the token. Otherwise, the connectors will encounter errors during operation.

pub fn get_application_id(&self) -> Result<String>

Gets the corresponding backend application identifier.

pub fn set_application_id(&mut self, app_id: String) -> Result<()>

Sets the corresponding backend application identifier.