Module cognitive_services_speech_sdk_rs::common

Package common contains the definitions for many of the shared objects and properties in the Speech SDK.

Structs

PropertyCollection
PropertyCollection is a class to retrieve or set a property value from a property collection.

Enums

CancellationErrorCode
CancellationErrorCode defines error code in case that CancellationReason is Error.

CancellationReason
CancellationReason defines the possible reasons a recognition result might be canceled.

OutputFormat
OutputFormat Defines output formats

ProfanityOption
ProfanityOption defines the profanity option.

PropertyId
PropertyID defines speech property ids.

ResultReason
ResultReason specifies the possible reasons a recognition result might be generated.

ServicePropertyChannel
ServicePropertyChannel defines channels used to pass property settings to service.

SpeechSynthesisOutputFormat
SpeechSynthesisOutputFormat defines the possible speech synthesis output audio formats.

StreamStatus
StreamStatus defines the possible status of audio data stream.

SynthesisVoiceType
SynthesisVoiceType defines the type of a synthesis voice.