Module cognitive_services_speech_sdk_rs::ffi

Package ffi contains bindings to underlying C API and thin abstraction for managing C handles.

Re-exports

pub use self::_ParticipantChangedReason as ParticipantChangedReason;

Structs

AudioProcessingOptions_MicrophoneArrayGeometry

AudioProcessingOptions_MicrophoneCoordinates

SmartHandle

Wrapper struct around underlying native handles Apart from handle it wraps release function that is automatically called when SmartHandle is dropped (see Drop trait implementation).

__darwin_pthread_handler_rec

_azac_empty

_opaque_pthread_attr_t

_opaque_pthread_cond_t

_opaque_pthread_condattr_t

_opaque_pthread_mutex_t

_opaque_pthread_mutexattr_t

_opaque_pthread_once_t

_opaque_pthread_rwlock_t

_opaque_pthread_rwlockattr_t

_opaque_pthread_t

codec_c_interface

imaxdiv_t

Constants

AUDIO_INPUT_PROCESSING_DISABLE_DEREVERBERATION

AUDIO_INPUT_PROCESSING_DISABLE_ECHO_CANCELLATION

AUDIO_INPUT_PROCESSING_DISABLE_GAIN_CONTROL

AUDIO_INPUT_PROCESSING_DISABLE_NOISE_SUPPRESSION

AUDIO_INPUT_PROCESSING_ENABLE_DEFAULT

AUDIO_INPUT_PROCESSING_ENABLE_V2

AUDIO_INPUT_PROCESSING_ENABLE_VOICE_ACTIVITY_DETECTION

AUDIO_INPUT_PROCESSING_NONE

AZAC_CONFIG_TRACE_INCLUDE_DBG_WITH_ALL

AZAC_ERR_NONE

AudioProcessingOptions_MicrophoneArrayType_AudioProcessingOptions_MicrophoneArrayType_Linear

AudioProcessingOptions_MicrophoneArrayType_AudioProcessingOptions_MicrophoneArrayType_Planar

AudioProcessingOptions_PresetMicrophoneArrayGeometry_AudioProcessingOptions_PresetMicrophoneArrayGeometry_Circular4

AudioProcessingOptions_PresetMicrophoneArrayGeometry_AudioProcessingOptions_PresetMicrophoneArrayGeometry_Circular7

AudioProcessingOptions_PresetMicrophoneArrayGeometry_AudioProcessingOptions_PresetMicrophoneArrayGeometry_Custom

AudioProcessingOptions_PresetMicrophoneArrayGeometry_AudioProcessingOptions_PresetMicrophoneArrayGeometry_Linear2

AudioProcessingOptions_PresetMicrophoneArrayGeometry_AudioProcessingOptions_PresetMicrophoneArrayGeometry_Linear4

AudioProcessingOptions_PresetMicrophoneArrayGeometry_AudioProcessingOptions_PresetMicrophoneArrayGeometry_Mono

AudioProcessingOptions_PresetMicrophoneArrayGeometry_AudioProcessingOptions_PresetMicrophoneArrayGeometry_Uninitialized

AudioProcessingOptions_SpeakerReferenceChannel_AudioProcessingOptions_SpeakerReferenceChannel_LastChannel

AudioProcessingOptions_SpeakerReferenceChannel_AudioProcessingOptions_SpeakerReferenceChannel_None

Audio_Stream_Container_Format_StreamFormat_Alaw

Audio_Stream_Container_Format_StreamFormat_Amrnb

Audio_Stream_Container_Format_StreamFormat_Amrwb

Audio_Stream_Container_Format_StreamFormat_Any

Audio_Stream_Container_Format_StreamFormat_Flac

Audio_Stream_Container_Format_StreamFormat_Mp3

Audio_Stream_Container_Format_StreamFormat_Mulaw

Audio_Stream_Container_Format_StreamFormat_Ogg_Opus

Audio_Stream_Wave_Format_StreamWaveFormat_ALAW

Audio_Stream_Wave_Format_StreamWaveFormat_MULAW

Audio_Stream_Wave_Format_StreamWaveFormat_PCM

GrammarList_RecognitionFactorScope_PartialPhrase

INT8_MAX

INT8_MIN

INT16_MAX

INT16_MIN

INT32_MAX

INT32_MIN

INT64_MAX

INT64_MIN

INTPTR_MAX

INTPTR_MIN

INT_FAST8_MAX

INT_FAST8_MIN

INT_FAST16_MAX

INT_FAST16_MIN

INT_FAST32_MAX

INT_FAST32_MIN

INT_FAST64_MAX

INT_FAST64_MIN

INT_LEAST8_MAX

INT_LEAST8_MIN

INT_LEAST16_MAX

INT_LEAST16_MIN

INT_LEAST32_MAX

INT_LEAST32_MIN

INT_LEAST64_MAX

INT_LEAST64_MIN

MAC_OS_VERSION_11_0

MAC_OS_VERSION_11_1

MAC_OS_VERSION_11_3

MAC_OS_VERSION_11_4

MAC_OS_VERSION_11_5

MAC_OS_VERSION_11_6

MAC_OS_VERSION_12_0

MAC_OS_VERSION_12_1

MAC_OS_VERSION_12_2

MAC_OS_VERSION_12_3

MAC_OS_VERSION_12_4

MAC_OS_VERSION_12_5

MAC_OS_VERSION_12_6

MAC_OS_VERSION_12_7

MAC_OS_VERSION_13_0

MAC_OS_VERSION_13_1

MAC_OS_VERSION_13_2

MAC_OS_VERSION_13_3

MAC_OS_VERSION_13_4

MAC_OS_VERSION_13_5

MAC_OS_VERSION_13_6

MAC_OS_VERSION_14_0

MAC_OS_VERSION_14_1

MAC_OS_VERSION_14_2

MAC_OS_VERSION_14_3

MAC_OS_VERSION_14_4

MAC_OS_X_VERSION_10_0

MAC_OS_X_VERSION_10_1

MAC_OS_X_VERSION_10_2

MAC_OS_X_VERSION_10_3

MAC_OS_X_VERSION_10_4

MAC_OS_X_VERSION_10_5

MAC_OS_X_VERSION_10_6

MAC_OS_X_VERSION_10_7

MAC_OS_X_VERSION_10_8

MAC_OS_X_VERSION_10_9

MAC_OS_X_VERSION_10_10

MAC_OS_X_VERSION_10_10_2

MAC_OS_X_VERSION_10_10_3

MAC_OS_X_VERSION_10_11

MAC_OS_X_VERSION_10_11_2

MAC_OS_X_VERSION_10_11_3

MAC_OS_X_VERSION_10_11_4

MAC_OS_X_VERSION_10_12

MAC_OS_X_VERSION_10_12_1

MAC_OS_X_VERSION_10_12_2

MAC_OS_X_VERSION_10_12_4

MAC_OS_X_VERSION_10_13

MAC_OS_X_VERSION_10_13_1

MAC_OS_X_VERSION_10_13_2

MAC_OS_X_VERSION_10_13_4

MAC_OS_X_VERSION_10_14

MAC_OS_X_VERSION_10_14_1

MAC_OS_X_VERSION_10_14_4

MAC_OS_X_VERSION_10_14_5

MAC_OS_X_VERSION_10_14_6

MAC_OS_X_VERSION_10_15

MAC_OS_X_VERSION_10_15_1

MAC_OS_X_VERSION_10_15_4

MAC_OS_X_VERSION_10_16

NULL_C_STR_PTR

PRIX8

PRIX16

PRIX32

PRIX64

PRIXFAST8

PRIXFAST16

PRIXFAST32

PRIXFAST64

PRIXLEAST8

PRIXLEAST16

PRIXLEAST32

PRIXLEAST64

PRIXMAX

PRIXPTR

PRId8

PRId16

PRId32

PRId64

PRIdFAST8

PRIdFAST16

PRIdFAST32

PRIdFAST64

PRIdLEAST8

PRIdLEAST16

PRIdLEAST32

PRIdLEAST64

PRIdMAX

PRIdPTR

PRIi8

PRIi16

PRIi32

PRIi64

PRIiFAST8

PRIiFAST16

PRIiFAST32

PRIiFAST64

PRIiLEAST8

PRIiLEAST16

PRIiLEAST32

PRIiLEAST64

PRIiMAX

PRIiPTR

PRIo8

PRIo16

PRIo32

PRIo64

PRIoFAST8

PRIoFAST16

PRIoFAST32

PRIoFAST64

PRIoLEAST8

PRIoLEAST16

PRIoLEAST32

PRIoLEAST64

PRIoMAX

PRIoPTR

PRIu8

PRIu16

PRIu32

PRIu64

PRIuFAST8

PRIuFAST16

PRIuFAST32

PRIuFAST64

PRIuLEAST8

PRIuLEAST16

PRIuLEAST32

PRIuLEAST64

PRIuMAX

PRIuPTR

PRIx8

PRIx16

PRIx32

PRIx64

PRIxFAST8

PRIxFAST16

PRIxFAST32

PRIxFAST64

PRIxLEAST8

PRIxLEAST16

PRIxLEAST32

PRIxLEAST64

PRIxMAX

PRIxPTR

Pronunciation_Assessment_Grading_System_PronunciationAssessmentGradingSystem_FivePoint

Pronunciation_Assessment_Grading_System_PronunciationAssessmentGradingSystem_HundredMark

Pronunciation_Assessment_Granularity_PronunciationAssessmentGranularity_FullText

Pronunciation_Assessment_Granularity_PronunciationAssessmentGranularity_Phoneme

Pronunciation_Assessment_Granularity_PronunciationAssessmentGranularity_Word

PropertyId_AudioConfig_AudioSource

PropertyId_AudioConfig_BitsPerSampleForCapture

PropertyId_AudioConfig_DeviceNameForCapture

PropertyId_AudioConfig_DeviceNameForRender

PropertyId_AudioConfig_NumberOfChannelsForCapture

PropertyId_AudioConfig_PlaybackBufferLengthInMs

PropertyId_AudioConfig_SampleRateForCapture

PropertyId_CancellationDetails_Reason

PropertyId_CancellationDetails_ReasonDetailedText

PropertyId_CancellationDetails_ReasonText

PropertyId_Conversation_ApplicationId

PropertyId_Conversation_Conversation_Id

PropertyId_Conversation_Custom_Voice_Deployment_Ids

PropertyId_Conversation_DialogType

PropertyId_Conversation_From_Id

PropertyId_Conversation_Initial_Silence_Timeout

PropertyId_Conversation_ParticipantId

PropertyId_Conversation_Speech_Activity_Template

PropertyId_DataBuffer_TimeStamp

PropertyId_DataBuffer_UserId

PropertyId_EmbeddedSpeech_EnablePerformanceMetrics

PropertyId_KeywordRecognition_ModelKey

PropertyId_KeywordRecognition_ModelName

PropertyId_LanguageUnderstandingServiceResponse_JsonResult

PropertyId_PronunciationAssessment_ContentTopic

PropertyId_PronunciationAssessment_EnableMiscue

PropertyId_PronunciationAssessment_EnableProsodyAssessment

PropertyId_PronunciationAssessment_GradingSystem

PropertyId_PronunciationAssessment_Granularity

PropertyId_PronunciationAssessment_Json

PropertyId_PronunciationAssessment_NBestPhonemeCount

PropertyId_PronunciationAssessment_Params

PropertyId_PronunciationAssessment_PhonemeAlphabet

PropertyId_PronunciationAssessment_ReferenceText

PropertyId_SpeakerRecognition_Api_Version

PropertyId_SpeechServiceAuthorization_Token

PropertyId_SpeechServiceAuthorization_Type

PropertyId_SpeechServiceConnection_AutoDetectSourceLanguageResult

PropertyId_SpeechServiceConnection_AutoDetectSourceLanguages

PropertyId_SpeechServiceConnection_EnableAudioLogging

PropertyId_SpeechServiceConnection_EndSilenceTimeoutMs

PropertyId_SpeechServiceConnection_Endpoint

PropertyId_SpeechServiceConnection_EndpointId

PropertyId_SpeechServiceConnection_Host

PropertyId_SpeechServiceConnection_InitialSilenceTimeoutMs

PropertyId_SpeechServiceConnection_IntentRegion

PropertyId_SpeechServiceConnection_Key

PropertyId_SpeechServiceConnection_LanguageIdMode

PropertyId_SpeechServiceConnection_ProxyHostName

PropertyId_SpeechServiceConnection_ProxyPassword

PropertyId_SpeechServiceConnection_ProxyPort

PropertyId_SpeechServiceConnection_ProxyUserName

PropertyId_SpeechServiceConnection_RecoLanguage

PropertyId_SpeechServiceConnection_RecoMode

PropertyId_SpeechServiceConnection_RecoModelBackend

PropertyId_SpeechServiceConnection_RecoModelIniFile

PropertyId_SpeechServiceConnection_RecoModelKey

PropertyId_SpeechServiceConnection_RecoModelName

PropertyId_SpeechServiceConnection_Region

PropertyId_SpeechServiceConnection_SynthBackend

PropertyId_SpeechServiceConnection_SynthEnableCompressedAudioTransmission

PropertyId_SpeechServiceConnection_SynthLanguage

PropertyId_SpeechServiceConnection_SynthModelKey

PropertyId_SpeechServiceConnection_SynthOfflineDataPath

PropertyId_SpeechServiceConnection_SynthOfflineVoice

PropertyId_SpeechServiceConnection_SynthOutputFormat

PropertyId_SpeechServiceConnection_SynthVoice

PropertyId_SpeechServiceConnection_TranslationFeatures

PropertyId_SpeechServiceConnection_TranslationToLanguages

PropertyId_SpeechServiceConnection_TranslationVoice

PropertyId_SpeechServiceConnection_Url

PropertyId_SpeechServiceConnection_UserDefinedQueryParameters

PropertyId_SpeechServiceConnection_VoicesListEndpoint

PropertyId_SpeechServiceResponse_JsonErrorDetails

PropertyId_SpeechServiceResponse_JsonResult

PropertyId_SpeechServiceResponse_OutputFormatOption

PropertyId_SpeechServiceResponse_PostProcessingOption

PropertyId_SpeechServiceResponse_ProfanityOption

PropertyId_SpeechServiceResponse_RecognitionBackend

PropertyId_SpeechServiceResponse_RecognitionLatencyMs

PropertyId_SpeechServiceResponse_RequestDetailedResultTrueFalse

PropertyId_SpeechServiceResponse_RequestProfanityFilterTrueFalse

PropertyId_SpeechServiceResponse_RequestPunctuationBoundary

PropertyId_SpeechServiceResponse_RequestSentenceBoundary

PropertyId_SpeechServiceResponse_RequestSnr

PropertyId_SpeechServiceResponse_RequestWordBoundary

PropertyId_SpeechServiceResponse_RequestWordLevelTimestamps

PropertyId_SpeechServiceResponse_StablePartialResultThreshold

PropertyId_SpeechServiceResponse_SynthesisConnectionLatencyMs

PropertyId_SpeechServiceResponse_SynthesisEventsSyncToAudio

PropertyId_SpeechServiceResponse_SynthesisFinishLatencyMs

PropertyId_SpeechServiceResponse_SynthesisFirstByteLatencyMs

PropertyId_SpeechServiceResponse_SynthesisNetworkLatencyMs

PropertyId_SpeechServiceResponse_SynthesisServiceLatencyMs

PropertyId_SpeechServiceResponse_SynthesisUnderrunTimeMs

PropertyId_SpeechServiceResponse_TranslationRequestStablePartialResult

PropertyId_SpeechTranslation_ModelKey

PropertyId_SpeechTranslation_ModelName

PropertyId_Speech_LogFilename

PropertyId_Speech_SegmentationSilenceTimeoutMs

PropertyId_Speech_SessionId

RSIZE_MAX

Result_CancellationErrorCode_CancellationErrorCode_AuthenticationFailure

Result_CancellationErrorCode_CancellationErrorCode_BadRequest

Result_CancellationErrorCode_CancellationErrorCode_ConnectionFailure

Result_CancellationErrorCode_CancellationErrorCode_Forbidden

Result_CancellationErrorCode_CancellationErrorCode_NoError

Result_CancellationErrorCode_CancellationErrorCode_RuntimeError

Result_CancellationErrorCode_CancellationErrorCode_ServiceError

Result_CancellationErrorCode_CancellationErrorCode_ServiceTimeout

Result_CancellationErrorCode_CancellationErrorCode_ServiceUnavailable

Result_CancellationErrorCode_CancellationErrorCode_TooManyRequests

Result_CancellationReason_CancellationReason_EndOfStream

Result_CancellationReason_CancellationReason_Error

Result_CancellationReason_CancellationReason_UserCancelled

Result_NoMatchReason_NoMatchReason_EndSilenceTimeout

Result_NoMatchReason_NoMatchReason_InitialBabbleTimeout

Result_NoMatchReason_NoMatchReason_InitialSilenceTimeout

Result_NoMatchReason_NoMatchReason_KeywordNotRecognized

Result_NoMatchReason_NoMatchReason_NotRecognized

Result_Reason_ResultReason_Canceled

Result_Reason_ResultReason_NoMatch

Result_Reason_ResultReason_RecognizedIntent

Result_Reason_ResultReason_RecognizedKeyword

Result_Reason_ResultReason_RecognizedSpeech

Result_Reason_ResultReason_RecognizingIntent

Result_Reason_ResultReason_RecognizingKeyword

Result_Reason_ResultReason_RecognizingSpeech

Result_Reason_ResultReason_SynthesizingAudio

Result_Reason_ResultReason_SynthesizingAudioComplete

Result_Reason_ResultReason_SynthesizingAudioStart

Result_Reason_ResultReason_TranslatedSpeech

Result_Reason_ResultReason_TranslatingSpeech

SCNd8

SCNd16

SCNd32

SCNd64

SCNdFAST8

SCNdFAST16

SCNdFAST32

SCNdFAST64

SCNdLEAST8

SCNdLEAST16

SCNdLEAST32

SCNdLEAST64

SCNdMAX

SCNdPTR

SCNi8

SCNi16

SCNi32

SCNi64

SCNiFAST8

SCNiFAST16

SCNiFAST32

SCNiFAST64

SCNiLEAST8

SCNiLEAST16

SCNiLEAST32

SCNiLEAST64

SCNiMAX

SCNiPTR

SCNo8

SCNo16

SCNo32

SCNo64

SCNoFAST8

SCNoFAST16

SCNoFAST32

SCNoFAST64

SCNoLEAST8

SCNoLEAST16

SCNoLEAST32

SCNoLEAST64

SCNoMAX

SCNoPTR

SCNu8

SCNu16

SCNu32

SCNu64

SCNuFAST8

SCNuFAST16

SCNuFAST32

SCNuFAST64

SCNuLEAST8

SCNuLEAST16

SCNuLEAST32

SCNuLEAST64

SCNuMAX

SCNuPTR

SCNx8

SCNx16

SCNx32

SCNx64

SCNxFAST8

SCNxFAST16

SCNxFAST32

SCNxFAST64

SCNxLEAST8

SCNxLEAST16

SCNxLEAST32

SCNxLEAST64

SCNxMAX

SCNxPTR

SIG_ATOMIC_MAX

SIG_ATOMIC_MIN

SIZE_MAX

SPX_CONFIG_TRACE_INCLUDE_DBG_WITH_ALL

SPX_NOERROR

SpeechConfig_ProfanityOption_SpeechConfig_ProfanityMasked

SpeechConfig_ProfanityOption_SpeechConfig_ProfanityRaw

SpeechConfig_ProfanityOption_SpeechConfig_ProfanityRemoved

SpeechConfig_ServicePropertyChannel_SpeechConfig_ServicePropertyChannel_HttpHeader

SpeechConfig_ServicePropertyChannel_SpeechConfig_ServicePropertyChannel_UriQueryParameter

SpeechOutputFormat_SpeechOutputFormat_Detailed

SpeechOutputFormat_SpeechOutputFormat_Simple

SpeechSynthesis_BoundaryType_SpeechSynthesis_BoundaryType_Punctuation

SpeechSynthesis_BoundaryType_SpeechSynthesis_BoundaryType_Sentence

SpeechSynthesis_BoundaryType_SpeechSynthesis_BoundaryType_Word

Speech_Synthesis_Output_Format_SpeechSynthesisOutputFormat_AmrWb16000Hz

amr-wb-16000hz AMR-WB audio at 16kHz sampling rate. (Added in 1.24.0)

Speech_Synthesis_Output_Format_SpeechSynthesisOutputFormat_Audio16Khz16Bit32KbpsMonoOpus

Speech_Synthesis_Output_Format_SpeechSynthesisOutputFormat_Audio16Khz16KbpsMonoSiren

Speech_Synthesis_Output_Format_SpeechSynthesisOutputFormat_Audio16Khz32KBitRateMonoMp3

Speech_Synthesis_Output_Format_SpeechSynthesisOutputFormat_Audio16Khz64KBitRateMonoMp3

Speech_Synthesis_Output_Format_SpeechSynthesisOutputFormat_Audio16Khz128KBitRateMonoMp3

Speech_Synthesis_Output_Format_SpeechSynthesisOutputFormat_Audio24Khz16Bit24KbpsMonoOpus

Speech_Synthesis_Output_Format_SpeechSynthesisOutputFormat_Audio24Khz16Bit48KbpsMonoOpus

Speech_Synthesis_Output_Format_SpeechSynthesisOutputFormat_Audio24Khz48KBitRateMonoMp3

Speech_Synthesis_Output_Format_SpeechSynthesisOutputFormat_Audio24Khz96KBitRateMonoMp3

Speech_Synthesis_Output_Format_SpeechSynthesisOutputFormat_Audio24Khz160KBitRateMonoMp3

Speech_Synthesis_Output_Format_SpeechSynthesisOutputFormat_Audio48Khz96KBitRateMonoMp3

Speech_Synthesis_Output_Format_SpeechSynthesisOutputFormat_Audio48Khz192KBitRateMonoMp3

Speech_Synthesis_Output_Format_SpeechSynthesisOutputFormat_Ogg16khz16BitMonoOpus

Speech_Synthesis_Output_Format_SpeechSynthesisOutputFormat_Ogg24Khz16BitMonoOpus

Speech_Synthesis_Output_Format_SpeechSynthesisOutputFormat_Ogg48Khz16BitMonoOpus

Speech_Synthesis_Output_Format_SpeechSynthesisOutputFormat_Raw8Khz8BitMonoALaw

Speech_Synthesis_Output_Format_SpeechSynthesisOutputFormat_Raw8Khz8BitMonoMULaw

Speech_Synthesis_Output_Format_SpeechSynthesisOutputFormat_Raw8Khz16BitMonoPcm

Speech_Synthesis_Output_Format_SpeechSynthesisOutputFormat_Raw16Khz16BitMonoPcm

Speech_Synthesis_Output_Format_SpeechSynthesisOutputFormat_Raw16Khz16BitMonoTrueSilk

Speech_Synthesis_Output_Format_SpeechSynthesisOutputFormat_Raw24Khz16BitMonoPcm

Speech_Synthesis_Output_Format_SpeechSynthesisOutputFormat_Raw24Khz16BitMonoTrueSilk

Speech_Synthesis_Output_Format_SpeechSynthesisOutputFormat_Raw48Khz16BitMonoPcm

Speech_Synthesis_Output_Format_SpeechSynthesisOutputFormat_Raw22050Hz16BitMonoPcm

Speech_Synthesis_Output_Format_SpeechSynthesisOutputFormat_Raw44100Hz16BitMonoPcm

Speech_Synthesis_Output_Format_SpeechSynthesisOutputFormat_Riff8Khz8BitMonoALaw

Speech_Synthesis_Output_Format_SpeechSynthesisOutputFormat_Riff8Khz8BitMonoMULaw

Speech_Synthesis_Output_Format_SpeechSynthesisOutputFormat_Riff8Khz16BitMonoPcm

Speech_Synthesis_Output_Format_SpeechSynthesisOutputFormat_Riff16Khz16BitMonoPcm

Speech_Synthesis_Output_Format_SpeechSynthesisOutputFormat_Riff16Khz16KbpsMonoSiren

Speech_Synthesis_Output_Format_SpeechSynthesisOutputFormat_Riff24Khz16BitMonoPcm

Speech_Synthesis_Output_Format_SpeechSynthesisOutputFormat_Riff48Khz16BitMonoPcm

Speech_Synthesis_Output_Format_SpeechSynthesisOutputFormat_Riff22050Hz16BitMonoPcm

Speech_Synthesis_Output_Format_SpeechSynthesisOutputFormat_Riff44100Hz16BitMonoPcm

Speech_Synthesis_Output_Format_SpeechSynthesisOutputFormat_Webm16Khz16BitMonoOpus

Speech_Synthesis_Output_Format_SpeechSynthesisOutputFormat_Webm24Khz16Bit24KbpsMonoOpus

Speech_Synthesis_Output_Format_SpeechSynthesisOutputFormat_Webm24Khz16BitMonoOpus

Stream_Status_StreamStatus_AllData

Stream_Status_StreamStatus_Canceled

Stream_Status_StreamStatus_NoData

Stream_Status_StreamStatus_PartialData

Stream_Status_StreamStatus_Unknown

Synthesis_VoiceType_SynthesisVoiceType_OfflineNeural

Synthesis_VoiceType_SynthesisVoiceType_OfflineStandard

Synthesis_VoiceType_SynthesisVoiceType_OnlineNeural

Synthesis_VoiceType_SynthesisVoiceType_OnlineStandard

UINT8_MAX

UINT16_MAX

UINT32_MAX

UINT64_MAX

UINTPTR_MAX

UINT_FAST8_MAX

UINT_FAST16_MAX

UINT_FAST32_MAX

UINT_FAST64_MAX

UINT_LEAST8_MAX

UINT_LEAST16_MAX

UINT_LEAST32_MAX

UINT_LEAST64_MAX

WINT_MAX

WINT_MIN

_DARWIN_FEATURE_64_BIT_INODE

_DARWIN_FEATURE_ONLY_64_BIT_INODE

_DARWIN_FEATURE_ONLY_UNIX_CONFORMANCE

_DARWIN_FEATURE_ONLY_VERS_1050

_DARWIN_FEATURE_UNIX_CONFORMANCE

_FORTIFY_SOURCE

_ParticipantChangedReason_JoinedConversation

_ParticipantChangedReason_LeftConversation

_ParticipantChangedReason_Updated

__API_TO_BE_DEPRECATED

__API_TO_BE_DEPRECATED_DRIVERKIT

__API_TO_BE_DEPRECATED_IOS

__API_TO_BE_DEPRECATED_MACCATALYST

__API_TO_BE_DEPRECATED_MACOS

__API_TO_BE_DEPRECATED_TVOS

__API_TO_BE_DEPRECATED_VISIONOS

__API_TO_BE_DEPRECATED_WATCHOS

__AZAC_TRACE_LEVEL_ERROR

__AZAC_TRACE_LEVEL_INFO

__AZAC_TRACE_LEVEL_VERBOSE

__AZAC_TRACE_LEVEL_WARNING

__BRIDGEOS_2_0

__BRIDGEOS_3_0

__BRIDGEOS_3_1

__BRIDGEOS_3_4

__BRIDGEOS_4_0

__BRIDGEOS_4_1

__BRIDGEOS_5_0

__BRIDGEOS_5_1

__BRIDGEOS_5_3

__BRIDGEOS_6_0

__BRIDGEOS_6_2

__BRIDGEOS_6_4

__BRIDGEOS_6_5

__BRIDGEOS_6_6

__BRIDGEOS_7_0

__BRIDGEOS_7_1

__BRIDGEOS_7_2

__BRIDGEOS_7_3

__BRIDGEOS_7_4

__BRIDGEOS_7_6

__BRIDGEOS_8_0

__BRIDGEOS_8_1

__BRIDGEOS_8_2

__BRIDGEOS_8_3

__BRIDGEOS_8_4

__DARWIN_64_BIT_INO_T

__DARWIN_C_ANSI

__DARWIN_C_FULL

__DARWIN_C_LEVEL

__DARWIN_NON_CANCELABLE

__DARWIN_NO_LONG_LONG

__DARWIN_ONLY_64_BIT_INO_T

__DARWIN_ONLY_UNIX_CONFORMANCE

__DARWIN_ONLY_VERS_1050

__DARWIN_SUF_EXTSN

__DARWIN_UNIX03

__DARWIN_VERS_1050

__DARWIN_WCHAR_MIN

__DRIVERKIT_19_0

__DRIVERKIT_20_0

__DRIVERKIT_21_0

__DRIVERKIT_22_0

__DRIVERKIT_22_4

__DRIVERKIT_22_5

__DRIVERKIT_22_6

__DRIVERKIT_23_0

__DRIVERKIT_23_1

__DRIVERKIT_23_2

__DRIVERKIT_23_3

__DRIVERKIT_23_4

__ENABLE_LEGACY_MAC_AVAILABILITY

__GNUC_VA_LIST

__IPHONE_2_0

__IPHONE_2_1

__IPHONE_2_2

__IPHONE_3_0

__IPHONE_3_1

__IPHONE_3_2

__IPHONE_4_0

__IPHONE_4_1

__IPHONE_4_2

__IPHONE_4_3

__IPHONE_5_0

__IPHONE_5_1

__IPHONE_6_0

__IPHONE_6_1

__IPHONE_7_0

__IPHONE_7_1

__IPHONE_8_0

__IPHONE_8_1

__IPHONE_8_2

__IPHONE_8_3

__IPHONE_8_4

__IPHONE_9_0

__IPHONE_9_1

__IPHONE_9_2

__IPHONE_9_3

__IPHONE_10_0

__IPHONE_10_1

__IPHONE_10_2

__IPHONE_10_3

__IPHONE_11_0

__IPHONE_11_1

__IPHONE_11_2

__IPHONE_11_3

__IPHONE_11_4

__IPHONE_12_0

__IPHONE_12_1

__IPHONE_12_2

__IPHONE_12_3

__IPHONE_12_4

__IPHONE_13_0

__IPHONE_13_1

__IPHONE_13_2

__IPHONE_13_3

__IPHONE_13_4

__IPHONE_13_5

__IPHONE_13_6

__IPHONE_13_7

__IPHONE_14_0

__IPHONE_14_1

__IPHONE_14_2

__IPHONE_14_3

__IPHONE_14_4

__IPHONE_14_5

__IPHONE_14_6

__IPHONE_14_7

__IPHONE_14_8

__IPHONE_15_0

__IPHONE_15_1

__IPHONE_15_2

__IPHONE_15_3

__IPHONE_15_4

__IPHONE_15_5

__IPHONE_15_6

__IPHONE_15_7

__IPHONE_15_8

__IPHONE_16_0

__IPHONE_16_1

__IPHONE_16_2

__IPHONE_16_3

__IPHONE_16_4

__IPHONE_16_5

__IPHONE_16_6

__IPHONE_16_7

__IPHONE_17_0

__IPHONE_17_1

__IPHONE_17_2

__IPHONE_17_3

__IPHONE_17_4

__MAC_10_0

__MAC_10_1

__MAC_10_2

__MAC_10_3

__MAC_10_4

__MAC_10_5

__MAC_10_6

__MAC_10_7

__MAC_10_8

__MAC_10_9

__MAC_10_10

__MAC_10_10_2

__MAC_10_10_3

__MAC_10_11

__MAC_10_11_2

__MAC_10_11_3

__MAC_10_11_4

__MAC_10_12

__MAC_10_12_1

__MAC_10_12_2

__MAC_10_12_4

__MAC_10_13

__MAC_10_13_1

__MAC_10_13_2

__MAC_10_13_4

__MAC_10_14

__MAC_10_14_1

__MAC_10_14_4

__MAC_10_14_5

__MAC_10_14_6

__MAC_10_15

__MAC_10_15_1

__MAC_10_15_4

__MAC_10_16

__MAC_11_0

__MAC_11_1

__MAC_11_3

__MAC_11_4

__MAC_11_5

__MAC_11_6

__MAC_12_0

__MAC_12_1

__MAC_12_2

__MAC_12_3

__MAC_12_4

__MAC_12_5

__MAC_12_6

__MAC_12_7

__MAC_13_0

__MAC_13_1

__MAC_13_2

__MAC_13_3

__MAC_13_4

__MAC_13_5

__MAC_13_6

__MAC_14_0

__MAC_14_1

__MAC_14_2

__MAC_14_3

__MAC_14_4

__MAC_OS_X_VERSION_MAX_ALLOWED

__PRI_8_LENGTH_MODIFIER__

__PRI_64_LENGTH_MODIFIER__

__PRI_MAX_LENGTH_MODIFIER__

__PTHREAD_ATTR_SIZE__

__PTHREAD_CONDATTR_SIZE__

__PTHREAD_COND_SIZE__

__PTHREAD_MUTEXATTR_SIZE__

__PTHREAD_MUTEX_SIZE__

__PTHREAD_ONCE_SIZE__

__PTHREAD_RWLOCKATTR_SIZE__

__PTHREAD_RWLOCK_SIZE__

__PTHREAD_SIZE__

__SCN_64_LENGTH_MODIFIER__

__SCN_MAX_LENGTH_MODIFIER__

__SPX_TRACE_LEVEL_ERROR

__SPX_TRACE_LEVEL_INFO

__SPX_TRACE_LEVEL_VERBOSE

__SPX_TRACE_LEVEL_WARNING

__STDC_WANT_LIB_EXT1__

__TVOS_9_0

__TVOS_9_1

__TVOS_9_2

__TVOS_10_0

__TVOS_10_0_1

__TVOS_10_1

__TVOS_10_2

__TVOS_11_0

__TVOS_11_1

__TVOS_11_2

__TVOS_11_3

__TVOS_11_4

__TVOS_12_0

__TVOS_12_1

__TVOS_12_2

__TVOS_12_3

__TVOS_12_4

__TVOS_13_0

__TVOS_13_2

__TVOS_13_3

__TVOS_13_4

__TVOS_14_0

__TVOS_14_1

__TVOS_14_2

__TVOS_14_3

__TVOS_14_5

__TVOS_14_6

__TVOS_14_7

__TVOS_15_0

__TVOS_15_1

__TVOS_15_2

__TVOS_15_3

__TVOS_15_4

__TVOS_15_5

__TVOS_15_6

__TVOS_16_0

__TVOS_16_1

__TVOS_16_2

__TVOS_16_3

__TVOS_16_4

__TVOS_16_5

__TVOS_16_6

__TVOS_17_0

__TVOS_17_1

__TVOS_17_2

__TVOS_17_3

__TVOS_17_4

__VISIONOS_1_0

__VISIONOS_1_1

__WATCHOS_1_0

__WATCHOS_2_0

__WATCHOS_2_1

__WATCHOS_2_2

__WATCHOS_3_0

__WATCHOS_3_1

__WATCHOS_3_1_1

__WATCHOS_3_2

__WATCHOS_4_0

__WATCHOS_4_1

__WATCHOS_4_2

__WATCHOS_4_3

__WATCHOS_5_0

__WATCHOS_5_1

__WATCHOS_5_2

__WATCHOS_5_3

__WATCHOS_6_0

__WATCHOS_6_1

__WATCHOS_6_2

__WATCHOS_7_0

__WATCHOS_7_1

__WATCHOS_7_2

__WATCHOS_7_3

__WATCHOS_7_4

__WATCHOS_7_5

__WATCHOS_7_6

__WATCHOS_8_0

__WATCHOS_8_1

__WATCHOS_8_3

__WATCHOS_8_4

__WATCHOS_8_5

__WATCHOS_8_6

__WATCHOS_8_7

__WATCHOS_8_8

__WATCHOS_9_0

__WATCHOS_9_1

__WATCHOS_9_2

__WATCHOS_9_3

__WATCHOS_9_4

__WATCHOS_9_5

__WATCHOS_9_6

__WATCHOS_10_0

__WATCHOS_10_1

__WATCHOS_10_2

__WATCHOS_10_3

__WATCHOS_10_4

__WORDSIZE

__bool_true_false_are_defined

__has_ptrcheck

__has_safe_buffers

false_

true_

Statics

CODECCREATEEXPORTNAME

Functions

add_source_lang_config_to_auto_detect_source_lang_config⚠

audio_config_create_audio_input_from_a_microphone⚠

audio_config_create_audio_input_from_default_microphone⚠

audio_config_create_audio_input_from_stream⚠

audio_config_create_audio_input_from_wav_file_name⚠

audio_config_create_audio_output_from_a_speaker⚠

audio_config_create_audio_output_from_default_speaker⚠

audio_config_create_audio_output_from_stream⚠

audio_config_create_audio_output_from_wav_file_name⚠

audio_config_create_pull_audio_input_stream⚠

audio_config_create_push_audio_input_stream⚠

audio_config_get_audio_processing_options⚠

audio_config_get_property_bag⚠

audio_config_is_handle_valid⚠

audio_config_release⚠

audio_config_set_audio_processing_options⚠

audio_data_stream_can_read_data⚠

audio_data_stream_can_read_data_from_position⚠

audio_data_stream_create_from_file⚠

audio_data_stream_create_from_keyword_result⚠

audio_data_stream_create_from_result⚠

audio_data_stream_detach_input⚠

audio_data_stream_get_canceled_error_code⚠

audio_data_stream_get_position⚠

audio_data_stream_get_property_bag⚠

audio_data_stream_get_reason_canceled⚠

audio_data_stream_get_status⚠

audio_data_stream_is_handle_valid⚠

audio_data_stream_read⚠

audio_data_stream_read_from_position⚠

audio_data_stream_release⚠

audio_data_stream_save_to_wave_file⚠

audio_data_stream_set_position⚠

audio_processing_options_create⚠

audio_processing_options_create_from_microphone_array_geometry⚠

audio_processing_options_create_from_preset_microphone_array_geometry⚠

audio_processing_options_get_audio_processing_flags⚠

audio_processing_options_get_beamforming_end_angle⚠

audio_processing_options_get_beamforming_start_angle⚠

audio_processing_options_get_microphone_array_type⚠

audio_processing_options_get_microphone_coordinates⚠

audio_processing_options_get_microphone_count⚠

audio_processing_options_get_preset_microphone_array_geometry⚠

audio_processing_options_get_property_bag⚠

audio_processing_options_get_speaker_reference_channel⚠

audio_processing_options_is_handle_valid⚠

audio_processing_options_release⚠

audio_stream_create_pull_audio_input_stream⚠

audio_stream_create_pull_audio_output_stream⚠

audio_stream_create_push_audio_input_stream⚠

audio_stream_create_push_audio_output_stream⚠

audio_stream_format_create_from_compressed_format⚠

audio_stream_format_create_from_default_input⚠

audio_stream_format_create_from_default_output⚠

audio_stream_format_create_from_waveformat⚠

audio_stream_format_create_from_waveformat_pcm⚠

audio_stream_format_is_handle_valid⚠

audio_stream_format_release⚠

audio_stream_is_handle_valid⚠

audio_stream_release⚠

auto_detect_source_lang_config_get_property_bag⚠

auto_detect_source_lang_config_is_handle_valid⚠

auto_detect_source_lang_config_release⚠

bot_framework_config_from_authorization_token⚠

bot_framework_config_from_subscription⚠

class_language_model_assign_class⚠

class_language_model_from_storage_id⚠

codec_create⚠

Creates a codec object. This method needs to be exported from the dll @param codecid - codec id, can be null or empty if the library implements only one codec. @param codecContext - context to be used to call back to the caller @param property_read_func - function to read properties @return A codec object

connection_async_handle_release⚠

connection_close⚠

connection_connected_set_callback⚠

connection_disconnected_set_callback⚠

connection_from_conversation_translator⚠

connection_from_dialog_service_connector⚠

connection_from_recognizer⚠

connection_from_speech_synthesizer⚠

connection_get_property_bag⚠

connection_handle_is_valid⚠

connection_handle_release⚠

connection_message_get_data⚠

connection_message_get_data_size⚠

connection_message_get_property_bag⚠

connection_message_handle_is_valid⚠

connection_message_handle_release⚠

connection_message_received_event_get_message⚠

connection_message_received_event_handle_is_valid⚠

connection_message_received_event_handle_release⚠

connection_message_received_set_callback⚠

connection_open⚠

connection_send_message⚠

connection_send_message_async⚠

connection_send_message_data⚠

connection_send_message_data_async⚠

connection_send_message_wait_for⚠

connection_set_message_property⚠

conversation_create_from_config⚠

conversation_delete_conversation⚠

conversation_end_conversation⚠

conversation_get_conversation_id⚠

conversation_get_property_bag⚠

conversation_lock_conversation⚠

conversation_mute_all_participants⚠

conversation_mute_participant⚠

conversation_release_handle⚠

conversation_start_conversation⚠

conversation_transcription_result_get_speaker_id⚠

conversation_translator_canceled_set_callback⚠

conversation_translator_connection_connected_set_callback⚠

conversation_translator_connection_disconnected_set_callback⚠

conversation_translator_conversation_expiration_set_callback⚠

conversation_translator_create_from_config⚠

conversation_translator_event_get_expiration_time⚠

conversation_translator_event_get_participant_changed_at_index⚠

conversation_translator_event_get_participant_changed_reason⚠

conversation_translator_event_handle_is_valid⚠

conversation_translator_event_handle_release⚠

conversation_translator_get_property_bag⚠

conversation_translator_handle_is_valid⚠

conversation_translator_handle_release⚠

conversation_translator_join⚠

conversation_translator_join_with_id⚠

conversation_translator_leave⚠

conversation_translator_participant_get_avatar⚠

conversation_translator_participant_get_displayname⚠

conversation_translator_participant_get_id⚠

conversation_translator_participant_get_is_host⚠

conversation_translator_participant_get_is_muted⚠

conversation_translator_participant_get_is_using_tts⚠

conversation_translator_participants_changed_set_callback⚠

conversation_translator_result_get_original_lang⚠

conversation_translator_result_get_user_id⚠

conversation_translator_send_text_message⚠

conversation_translator_session_started_set_callback⚠

conversation_translator_session_stopped_set_callback⚠

conversation_translator_set_authorization_token⚠

conversation_translator_start_transcribing⚠

conversation_translator_stop_transcribing⚠

conversation_translator_text_message_recevied_set_callback⚠

conversation_translator_transcribed_set_callback⚠

conversation_translator_transcribing_set_callback⚠

conversation_unlock_conversation⚠

conversation_unmute_all_participants⚠

conversation_unmute_participant⚠

conversation_update_participant⚠

conversation_update_participant_by_user⚠

conversation_update_participant_by_user_id⚠

create_auto_detect_source_lang_config_from_languages⚠

create_auto_detect_source_lang_config_from_open_range⚠

create_auto_detect_source_lang_config_from_source_lang_config⚠

create_pronunciation_assessment_config⚠

create_pronunciation_assessment_config_from_json⚠

create_voice_profile⚠

create_voice_profile_client_from_config⚠

create_voice_profile_from_id_and_type⚠

custom_commands_config_from_authorization_token⚠

custom_commands_config_from_subscription⚠

delete_voice_profile⚠

diagnostics_eventsource_logmessage_set_callback⚠

diagnostics_eventsource_logmessage_set_filters⚠

diagnostics_free_string⚠

diagnostics_get_handle_count⚠

diagnostics_get_handle_info⚠

diagnostics_is_log_level_enabled⚠

diagnostics_log_apply_properties⚠

diagnostics_log_console_set_filters⚠

diagnostics_log_console_start_logging⚠

diagnostics_log_console_stop_logging⚠

diagnostics_log_format_message⚠

diagnostics_log_memory_dump⚠

diagnostics_log_memory_dump_on_exit⚠

diagnostics_log_memory_dump_to_stderr⚠

diagnostics_log_memory_get_line⚠

diagnostics_log_memory_get_line_num_newest⚠

diagnostics_log_memory_get_line_num_oldest⚠

diagnostics_log_memory_set_filters⚠

diagnostics_log_memory_start_logging⚠

diagnostics_log_memory_stop_logging⚠

diagnostics_log_start_logging⚠

diagnostics_log_stop_logging⚠

diagnostics_log_trace_message⚠

diagnostics_log_trace_message2⚠

diagnostics_log_trace_string⚠

diagnostics_logmessage_set_callback⚠

diagnostics_logmessage_set_filters⚠

diagnostics_set_log_level⚠

dialog_service_connector_activity_received_event_get_activity⚠

dialog_service_connector_activity_received_event_get_activity_size⚠

dialog_service_connector_activity_received_event_get_audio⚠

dialog_service_connector_activity_received_event_handle_is_valid⚠

dialog_service_connector_activity_received_event_has_audio⚠

dialog_service_connector_activity_received_event_release⚠

dialog_service_connector_activity_received_set_callback⚠

dialog_service_connector_async_handle_is_valid⚠

dialog_service_connector_async_handle_release⚠

dialog_service_connector_async_reco_result_handle_is_valid⚠

dialog_service_connector_async_reco_result_handle_release⚠

dialog_service_connector_async_string_handle_is_valid⚠

dialog_service_connector_async_string_handle_release⚠

dialog_service_connector_async_void_handle_is_valid⚠

dialog_service_connector_async_void_handle_release⚠

dialog_service_connector_canceled_set_callback⚠

dialog_service_connector_connect⚠

dialog_service_connector_connect_async⚠

dialog_service_connector_connect_async_wait_for⚠

dialog_service_connector_create_dialog_service_connector_from_config⚠

dialog_service_connector_disconnect⚠

dialog_service_connector_disconnect_async⚠

dialog_service_connector_disconnect_async_wait_for⚠

dialog_service_connector_get_property_bag⚠

dialog_service_connector_handle_is_valid⚠

dialog_service_connector_handle_release⚠

dialog_service_connector_listen_once⚠

dialog_service_connector_listen_once_async⚠

dialog_service_connector_listen_once_async_wait_for⚠

dialog_service_connector_recognized_get_result⚠

dialog_service_connector_recognized_set_callback⚠

dialog_service_connector_recognized_size⚠

dialog_service_connector_recognizing_set_callback⚠

dialog_service_connector_send_activity⚠

dialog_service_connector_send_activity_async⚠

dialog_service_connector_send_activity_async_wait_for⚠

dialog_service_connector_session_started_set_callback⚠

dialog_service_connector_session_stopped_set_callback⚠

dialog_service_connector_speech_end_detected_set_callback⚠

dialog_service_connector_speech_start_detected_set_callback⚠

dialog_service_connector_start_continuous_listening⚠

dialog_service_connector_start_continuous_listening_async⚠

dialog_service_connector_start_keyword_recognition⚠

dialog_service_connector_start_keyword_recognition_async⚠

dialog_service_connector_start_keyword_recognition_async_wait_for⚠

dialog_service_connector_stop_keyword_recognition⚠

dialog_service_connector_stop_keyword_recognition_async⚠

dialog_service_connector_stop_keyword_recognition_async_wait_for⚠

dialog_service_connector_stop_listening⚠

dialog_service_connector_stop_listening_async⚠

dialog_service_connector_turn_status_received_get_conversation_id⚠

dialog_service_connector_turn_status_received_get_conversation_id_size⚠

dialog_service_connector_turn_status_received_get_interaction_id⚠

dialog_service_connector_turn_status_received_get_interaction_id_size⚠

dialog_service_connector_turn_status_received_get_status⚠

dialog_service_connector_turn_status_received_handle_is_valid⚠

dialog_service_connector_turn_status_received_release⚠

dialog_service_connector_turn_status_received_set_callback⚠

embedded_speech_config_add_path⚠

embedded_speech_config_create⚠

embedded_speech_config_get_num_speech_reco_models⚠

embedded_speech_config_get_num_speech_translation_models⚠

embedded_speech_config_get_speech_reco_model⚠

embedded_speech_config_get_speech_translation_model⚠

enroll_voice_profile⚠

error_get_call_stack⚠

error_get_error_code⚠

error_get_message⚠

error_release⚠

get_activation_phrases⚠

get_profiles_json⚠

grammar_create_from_storage_id⚠

grammar_handle_is_valid⚠

grammar_handle_release⚠

grammar_list_add_grammar⚠

grammar_list_from_recognizer⚠

grammar_list_set_recognition_factor⚠

grammar_phrase_create_from_text⚠

grammar_phrase_handle_is_valid⚠

grammar_phrase_handle_release⚠

hybrid_speech_config_create⚠

imaxabs⚠

imaxdiv⚠

intent_recognizer_add_conversational_language_understanding_model⚠

intent_recognizer_add_intent⚠

intent_recognizer_add_intent_with_model_id⚠

intent_recognizer_clear_language_models⚠

intent_recognizer_import_pattern_matching_model⚠

intent_recognizer_recognize_text_once⚠

intent_result_get_intent_id⚠

intent_trigger_create_from_language_understanding_model⚠

intent_trigger_create_from_phrase⚠

intent_trigger_handle_is_valid⚠

intent_trigger_handle_release⚠

keyword_recognition_model_add_user_defined_wake_word⚠

keyword_recognition_model_create_from_config⚠

keyword_recognition_model_create_from_file⚠

keyword_recognition_model_handle_is_valid⚠

keyword_recognition_model_handle_release⚠

language_understanding_model__handle_release⚠

language_understanding_model_create_from_app_id⚠

language_understanding_model_create_from_subscription⚠

language_understanding_model_create_from_uri⚠

language_understanding_model_get_model_id⚠

language_understanding_model_handle_is_valid⚠

meeting_create_from_config⚠

meeting_delete_meeting⚠

meeting_end_meeting⚠

meeting_get_meeting_id⚠

meeting_get_property_bag⚠

meeting_lock_meeting⚠

meeting_mute_all_participants⚠

meeting_mute_participant⚠

meeting_release_handle⚠

meeting_start_meeting⚠

meeting_transcription_result_get_user_id⚠

meeting_transcription_result_get_utterance_id⚠

meeting_unlock_meeting⚠

meeting_unmute_all_participants⚠

meeting_unmute_participant⚠

meeting_update_participant⚠

meeting_update_participant_by_user⚠

meeting_update_participant_by_user_id⚠

participant_create_handle⚠

participant_get_property_bag⚠

participant_release_handle⚠

participant_set_preferred_langugage⚠

participant_set_voice_signature⚠

pattern_matching_model_add_entity⚠

pattern_matching_model_add_intent⚠

pattern_matching_model_create⚠

pattern_matching_model_create_from_id⚠

pattern_matching_model_handle_is_valid⚠

phrase_list_grammar_add_phrase⚠

phrase_list_grammar_clear⚠

phrase_list_grammar_from_recognizer_by_name⚠

pronunciation_assessment_config_apply_to_recognizer⚠

pronunciation_assessment_config_get_property_bag⚠

pronunciation_assessment_config_is_handle_valid⚠

pronunciation_assessment_config_release⚠

pronunciation_assessment_config_to_json⚠

property_bag_copy⚠

property_bag_create⚠

property_bag_free_string⚠

property_bag_get_string⚠

property_bag_is_valid⚠

property_bag_release⚠

property_bag_set_string⚠

pull_audio_input_stream_set_callbacks⚠

pull_audio_input_stream_set_getproperty_callback⚠

pull_audio_output_stream_read⚠

push_audio_input_stream_close⚠

push_audio_input_stream_set_property_by_id⚠

push_audio_input_stream_set_property_by_name⚠

push_audio_input_stream_write⚠

push_audio_output_stream_set_callbacks⚠

recognizer_async_handle_is_valid⚠

recognizer_async_handle_release⚠

recognizer_canceled_set_callback⚠

recognizer_connection_event_get_property_bag⚠

recognizer_create_conversation_transcriber_from_auto_detect_source_lang_config⚠

recognizer_create_conversation_transcriber_from_config⚠

recognizer_create_conversation_transcriber_from_source_lang_config⚠

recognizer_create_intent_recognizer_from_config⚠

recognizer_create_keyword_recognizer_from_audio_config⚠

recognizer_create_meeting_transcriber_from_config⚠

recognizer_create_source_language_recognizer_from_auto_detect_source_lang_config⚠

recognizer_create_speaker_recognizer_from_config⚠

recognizer_create_speech_recognizer_from_auto_detect_source_lang_config⚠

recognizer_create_speech_recognizer_from_config⚠

recognizer_create_speech_recognizer_from_source_lang_config⚠

recognizer_create_translation_recognizer_from_auto_detect_source_lang_config⚠

recognizer_create_translation_recognizer_from_config⚠

recognizer_event_handle_is_valid⚠

recognizer_event_handle_release⚠

recognizer_get_property_bag⚠

recognizer_handle_is_valid⚠

recognizer_handle_release⚠

recognizer_join_conversation⚠

recognizer_join_meeting⚠

recognizer_leave_conversation⚠

recognizer_leave_meeting⚠

recognizer_recognition_event_get_offset⚠

recognizer_recognition_event_get_result⚠

recognizer_recognize_keyword_once⚠

recognizer_recognize_keyword_once_async⚠

recognizer_recognize_keyword_once_async_wait_for⚠

recognizer_recognize_once⚠

recognizer_recognize_once_async⚠

recognizer_recognize_once_async_wait_for⚠

recognizer_recognize_text_once_async⚠

recognizer_recognized_set_callback⚠

recognizer_recognizing_set_callback⚠

recognizer_result_handle_is_valid⚠

recognizer_result_handle_release⚠

recognizer_session_event_get_session_id⚠

recognizer_session_started_set_callback⚠

recognizer_session_stopped_set_callback⚠

recognizer_speech_end_detected_set_callback⚠

recognizer_speech_start_detected_set_callback⚠

recognizer_start_continuous_recognition⚠

recognizer_start_continuous_recognition_async⚠

recognizer_start_continuous_recognition_async_wait_for⚠

recognizer_start_keyword_recognition⚠

recognizer_start_keyword_recognition_async⚠

recognizer_start_keyword_recognition_async_wait_for⚠

recognizer_stop_continuous_recognition⚠

recognizer_stop_continuous_recognition_async⚠

recognizer_stop_continuous_recognition_async_wait_for⚠

recognizer_stop_keyword_recognition⚠

recognizer_stop_keyword_recognition_async⚠

recognizer_stop_keyword_recognition_async_wait_for⚠

reset_voice_profile⚠

result_get_canceled_error_code⚠

result_get_duration⚠

result_get_no_match_reason⚠

result_get_offset⚠

result_get_property_bag⚠

result_get_reason⚠

result_get_reason_canceled⚠

result_get_result_id⚠

result_get_text⚠

retrieve_enrollment_result⚠

session_from_recognizer⚠

session_get_property_bag⚠

session_handle_is_valid⚠

session_handle_release⚠

source_lang_config_from_language⚠

source_lang_config_from_language_and_endpointId⚠

source_lang_config_get_property_bag⚠

source_lang_config_is_handle_valid⚠

source_lang_config_release⚠

speaker_identification_model_add_profile⚠

speaker_identification_model_create⚠

speaker_identification_model_release_handle⚠

speaker_recognizer_get_property_bag⚠

speaker_recognizer_identify⚠

speaker_recognizer_release_handle⚠

speaker_recognizer_verify⚠

speaker_verification_model_create⚠

speaker_verification_model_release_handle⚠

speech_config_from_authorization_token⚠

speech_config_from_endpoint⚠

speech_config_from_host⚠

speech_config_from_subscription⚠

speech_config_get_property_bag⚠

speech_config_is_handle_valid⚠

speech_config_release⚠

speech_config_set_audio_output_format⚠

speech_config_set_profanity⚠

speech_config_set_service_property⚠

speech_recognition_model_get_locales⚠

speech_recognition_model_get_name⚠

speech_recognition_model_get_path⚠

speech_recognition_model_get_version⚠

speech_recognition_model_handle_release⚠

speech_synthesis_request_create⚠

speech_synthesis_request_finish⚠

speech_synthesis_request_get_property_bag⚠

speech_synthesis_request_handle_is_valid⚠

speech_synthesis_request_release⚠

speech_synthesis_request_send_text_piece⚠

speech_translation_config_add_target_language⚠

speech_translation_config_from_authorization_token⚠

speech_translation_config_from_endpoint⚠

speech_translation_config_from_host⚠

speech_translation_config_from_subscription⚠

speech_translation_config_remove_target_language⚠

speech_translation_model_get_name⚠

speech_translation_model_get_path⚠

speech_translation_model_get_source_languages⚠

speech_translation_model_get_target_languages⚠

speech_translation_model_get_version⚠

speech_translation_model_handle_release⚠

strtoimax⚠

strtoumax⚠

synth_result_get_audio_data⚠

synth_result_get_audio_format⚠

synth_result_get_audio_length_duration⚠

synth_result_get_canceled_error_code⚠

synth_result_get_property_bag⚠

synth_result_get_reason⚠

synth_result_get_reason_canceled⚠

synth_result_get_result_id⚠

synthesis_voices_result_get_property_bag⚠

synthesis_voices_result_get_reason⚠

synthesis_voices_result_get_result_id⚠

synthesis_voices_result_get_voice_info⚠

synthesis_voices_result_get_voice_num⚠

synthesizer_async_handle_is_valid⚠

synthesizer_async_handle_release⚠

synthesizer_bookmark_event_get_values⚠

synthesizer_bookmark_reached_set_callback⚠

synthesizer_canceled_set_callback⚠

synthesizer_completed_set_callback⚠

synthesizer_connection_connected_set_callback⚠

synthesizer_connection_disconnected_set_callback⚠

synthesizer_create_speech_synthesizer_from_auto_detect_source_lang_config⚠

synthesizer_create_speech_synthesizer_from_config⚠

synthesizer_event_get_result_id⚠

synthesizer_event_get_text⚠

synthesizer_event_handle_is_valid⚠

synthesizer_event_handle_release⚠

synthesizer_get_property_bag⚠

synthesizer_get_voices_list⚠

synthesizer_get_voices_list_async⚠

synthesizer_get_voices_list_async_wait_for⚠

synthesizer_handle_is_valid⚠

synthesizer_handle_release⚠

synthesizer_result_handle_is_valid⚠

synthesizer_result_handle_release⚠

synthesizer_speak_async_wait_for⚠

synthesizer_speak_request⚠

synthesizer_speak_request_async⚠

synthesizer_speak_ssml⚠

synthesizer_speak_ssml_async⚠

synthesizer_speak_text⚠

synthesizer_speak_text_async⚠

synthesizer_start_speaking_request⚠

synthesizer_start_speaking_ssml⚠

synthesizer_start_speaking_ssml_async⚠

synthesizer_start_speaking_text⚠

synthesizer_start_speaking_text_async⚠

synthesizer_started_set_callback⚠

synthesizer_stop_speaking⚠

synthesizer_stop_speaking_async⚠

synthesizer_stop_speaking_async_wait_for⚠

synthesizer_synthesis_event_get_result⚠

synthesizer_synthesizing_set_callback⚠

synthesizer_viseme_event_get_animation⚠

synthesizer_viseme_event_get_values⚠

synthesizer_viseme_received_set_callback⚠

synthesizer_word_boundary_event_get_values⚠

synthesizer_word_boundary_set_callback⚠

transcriber_get_participants_list⚠

translation_synthesis_result_get_audio_data⚠

translation_text_result_get_translation⚠

translation_text_result_get_translation_count⚠

translator_add_target_language⚠

translator_remove_target_language⚠

translator_synthesizing_audio_set_callback⚠

user_create_from_id⚠

user_get_id⚠

user_release_handle⚠

voice_info_get_local_name⚠

voice_info_get_locale⚠

voice_info_get_name⚠

voice_info_get_property_bag⚠

voice_info_get_short_name⚠

voice_info_get_style_list⚠

voice_info_get_voice_path⚠

voice_info_get_voice_type⚠

voice_info_handle_release⚠

voice_profile_client_get_property_bag⚠

voice_profile_client_release_handle⚠

voice_profile_get_id⚠

voice_profile_get_property_bag⚠

voice_profile_get_type⚠

voice_profile_release_handle⚠

wcstoimax⚠

wcstoumax⚠

Type Definitions

AUDIO_ENCODER_ONENCODEDDATA

The AUDIO_ENCODER_ONENCODEDDATA type represents an application-defined status callback function used to provide the encoded data. @param pBuffer audio data buffer. @param bytesToWrite The length of pBuffer in bytes. @param duration_100nanos The duration of the audio sample @param pContext A pointer to the application-defined callback context.

AZACHR

AZAC_HANDLE

AudioProcessingOptions_MicrophoneArrayType

AudioProcessingOptions_PresetMicrophoneArrayGeometry

AudioProcessingOptions_SpeakerReferenceChannel

Audio_Stream_Container_Format

Audio_Stream_Wave_Format

CONNECTION_CALLBACK_FUNC

CUSTOM_AUDIO_PULL_STREAM_CLOSE_CALLBACK

CUSTOM_AUDIO_PULL_STREAM_GET_PROPERTY_CALLBACK

CUSTOM_AUDIO_PULL_STREAM_READ_CALLBACK

CUSTOM_AUDIO_PUSH_STREAM_CLOSE_CALLBACK

CUSTOM_AUDIO_PUSH_STREAM_WRITE_CALLBACK

DIAGNOSTICS_CALLBACK_FUNC

DIAGNOSTICS_EVENTSOURCE_CALLBACK_FUNC

GrammarList_RecognitionFactorScope

PATTERN_MATCHING_MODEL_GET_STR_FROM_INDEX

PCODEC_CREATE_FUNC

PCONV_TRANS_CALLBACK

PRECOGNITION_CALLBACK_FUNC

PSESSION_CALLBACK_FUNC

PSYNTHESIS_CALLBACK_FUNC

PTRANSLATIONSYNTHESIS_CALLBACK_FUNC

Pronunciation_Assessment_Grading_System

Pronunciation_Assessment_Granularity

PropertyId

Result_CancellationErrorCode

Result_CancellationReason

Result_NoMatchReason

Result_Reason

SPXACTIVITYHANDLE

SPXACTIVITYJSONHANDLE

SPXASYNCHANDLE

SPXAUDIOCONFIGHANDLE

SPXAUDIOPROCESSINGOPTIONSHANDLE

SPXAUDIOSTREAMFORMATHANDLE

SPXAUDIOSTREAMHANDLE

SPXAUTODETECTSOURCELANGCONFIGHANDLE

SPXCODECCTYPE

SPXCONNECTIONHANDLE

SPXCONNECTIONMESSAGEHANDLE

SPXCONVERSATIONHANDLE

SPXCONVERSATIONTRANSLATORHANDLE

SPXEVENTHANDLE

SPXFACTORYHANDLE

SPXGRAMMARHANDLE

SPXHANDLE

SPXHR

SPXKEYWORDHANDLE

SPXLUMODELHANDLE

SPXMEETINGHANDLE

SPXPARTICIPANTHANDLE

SPXPHRASEHANDLE

SPXPRONUNCIATIONASSESSMENTCONFIGHANDLE

SPXPROPERTYBAGHANDLE

SPXRECOHANDLE

SPXREQUESTHANDLE

SPXRESULTHANDLE

SPXSESSIONHANDLE

SPXSIMODELHANDLE

SPXSOURCELANGCONFIGHANDLE

SPXSPEAKERIDHANDLE

SPXSPEECHCONFIGHANDLE

SPXSPEECHRECOMODELHANDLE

SPXSVMODELHANDLE

SPXSYNTHHANDLE

SPXTRIGGERHANDLE

SPXUSERHANDLE

SPXVOICEPROFILECLIENTHANDLE

SPXVOICEPROFILEHANDLE

SPX_CODEC_CLIENT_GET_PROPERTY

The SPX_CODEC_CLIENT_GET_PROPERTY represents the function reading a property value @param id Property id. @param buffer caller provided buffer to receive the value of the property @param buffersize buffer size. If buffer is passed as null it will return the required buffer size. @param codecContext A pointer to caller data provided through the codec_create call. @return A return code or zero if successful.

SpeechConfig_ProfanityOption

SpeechConfig_ServicePropertyChannel

SpeechOutputFormat

SpeechSynthesis_BoundaryType

Speech_Synthesis_Output_Format

Stream_Status

Synthesis_VoiceType

_ParticipantChangedReason

__builtin_va_list

__darwin_blkcnt_t

__darwin_blksize_t

__darwin_clock_t

__darwin_ct_rune_t

__darwin_dev_t

__darwin_fsblkcnt_t

__darwin_fsfilcnt_t

__darwin_gid_t

__darwin_id_t

__darwin_ino64_t

__darwin_ino_t

__darwin_intptr_t

__darwin_mach_port_name_t

__darwin_mach_port_t

__darwin_mbstate_t

__darwin_mode_t

__darwin_natural_t

__darwin_nl_item

__darwin_off_t

__darwin_pid_t

__darwin_pthread_attr_t

__darwin_pthread_cond_t

__darwin_pthread_condattr_t

__darwin_pthread_key_t

__darwin_pthread_mutex_t

__darwin_pthread_mutexattr_t

__darwin_pthread_once_t

__darwin_pthread_rwlock_t

__darwin_pthread_rwlockattr_t

__darwin_pthread_t

__darwin_ptrdiff_t

__darwin_rune_t

__darwin_sigset_t

__darwin_size_t

__darwin_socklen_t

__darwin_ssize_t

__darwin_suseconds_t

__darwin_time_t

__darwin_uid_t

__darwin_useconds_t

__darwin_uuid_string_t

__darwin_uuid_t

__darwin_va_list

__darwin_wchar_t

__darwin_wctrans_t

__darwin_wctype_t

__darwin_wint_t

__gnuc_va_list

__int8_t

__int16_t

__int32_t

__int64_t

__uint8_t

__uint16_t

__uint32_t

__uint64_t

_azachandle

codec_c_interface_P

const_char_ptr

int_fast8_t

int_fast16_t

int_fast32_t

int_fast64_t

int_least8_t

int_least16_t

int_least32_t

int_least64_t

intmax_t

max_align_t

register_t

size_t

syscall_arg_t

u_int8_t

u_int16_t

u_int32_t

u_int64_t

uint_fast8_t

uint_fast16_t

uint_fast32_t

uint_fast64_t

uint_least8_t

uint_least16_t

uint_least32_t

uint_least64_t

uintmax_t

user_addr_t

user_long_t

user_off_t

user_size_t

user_ssize_t

user_time_t

user_ulong_t

va_list

wchar_t

Unions

__mbstate_t