Module cognitive_services_speech_sdk_rs::speech

Package speech provides functionality for speech recognizers along with their related configuration and event objects.

Structs

AudioDataStream
AudioDataStream represents audio data retrieved either from file or result of speech synthesis. Represents convenient option for manipulating and storing of syntehtized audio data Added in version 1.17.0

AutoDetectSourceLanguageConfig
AutoDetectSourceLanguageConfig defines auto detection source configuration

CancellationDetails
CancellationDetails contains detailed information about why a result was canceled. Added in version 1.17.0

KeywordRecognitionModel
KeywordRecognitionModel represents the keyword recognition model used with start_keyword_recognition_async methods.

RecognitionEvent
Recognition event extending SessionEvent passed into callbacks set_speech_start_detected_cb and set_speech_end_detected_cb.

SessionEvent
Base SpeechRecognizer event passed into callbacks set_session_started_cb and set_session_stopped_cb.

SourceLanguageConfig

SpeechConfig
SpeechConfig is the class that defines configurations for speech / intent recognition, or speech synthesis.

SpeechRecognitionCanceledEvent
Recognition event extending SpeechRecognitionEvent passed into callback set_canceled_cb.

SpeechRecognitionEvent
Recognition event extending RecognitionEvent passed into callbacks set_recognizing_cb and set_recognized_cb.

SpeechRecognitionResult
Represents speech recognition result contained within callback event SpeechRecognitionEvent.

SpeechRecognizer
SpeechRecognizer struct holds functionality for speech-to-text recognition.

SpeechSynthesisBookmarkEvent
Event passed into speech synthetizer's callback set_synthesizer_bookmark_cb.

SpeechSynthesisEvent
Event passed into speech synthetizer callbacks.

SpeechSynthesisResult
Represents speech synthetis result contained in SpeechSynthesisEvent callback event.

SpeechSynthesisVisemeEvent
Event passed into speech synthetizer's callback set_synthesizer_viseme_cb.

SpeechSynthesisWordBoundaryEvent
Event passed into speech synthetizer's callback set_synthesizer_word_boundary_cb.

SpeechSynthesizer
SpeechSynthesizer struct holds functionality for text-to-speech synthesis.

SynthesisVoicesResult

VoiceInfo