Struct cognitive_services_speech_sdk_rs::speech::SpeechRecognizer

pub struct SpeechRecognizer { /* fields omitted */ }

SpeechRecognizer struct holds functionality for speech-to-text recognition.

Implementations

impl SpeechRecognizer

pub fn from_config(
speech_config: SpeechConfig,
audio_config: AudioConfig
) -> Result<SpeechRecognizer>

pub fn from_auto_detect_source_lang_config(
speech_config: SpeechConfig,
audio_config: AudioConfig,
lang_config: AutoDetectSourceLanguageConfig
) -> Result<SpeechRecognizer>

pub fn from_source_lang_config(
speech_config: SpeechConfig,
audio_config: AudioConfig,
source_lang_config: SourceLanguageConfig
) -> Result<SpeechRecognizer>

pub fn from_source_lang(
speech_config: SpeechConfig,
audio_config: AudioConfig,
source_lang: &str
) -> Result<SpeechRecognizer>

pub fn set_session_started_cb<F>(&mut self, f: F) -> Result<()> where
F: Fn(SessionEvent) + 'static + Send,

pub fn set_session_stopped_cb<F>(&mut self, f: F) -> Result<()> where
F: Fn(SessionEvent) + 'static + Send,

pub fn set_speech_start_detected_cb<F>(&mut self, f: F) -> Result<()> where
F: Fn(RecognitionEvent) + 'static + Send,

pub fn set_speech_end_detected_cb<F>(&mut self, f: F) -> Result<()> where
F: Fn(RecognitionEvent) + 'static + Send,

pub fn set_canceled_cb<F>(&mut self, f: F) -> Result<()> where
F: Fn(SpeechRecognitionCanceledEvent) + 'static + Send,

Canceled signals for events containing canceled recognition results (indicating a recognition attempt that was canceled as a result or a direct cancellation request or, alternatively, a transport or protocol failure).

pub fn set_recognizing_cb<F>(&mut self, f: F) -> Result<()> where
F: Fn(SpeechRecognitionEvent) + 'static + Send,

pub fn set_recognized_cb<F>(&mut self, f: F) -> Result<()> where
F: Fn(SpeechRecognitionEvent) + 'static + Send,

pub async fn recognize_once_async<'_>(
&'_ mut self
) -> Result<SpeechRecognitionResult>

Starts speech recognition, and returns after a single utterance is recognized. The end of a single utterance is determined by listening for silence at the end or until a maximum of 15 seconds of audio is processed. The task returns the recognition text as result. Note: Since RecognizeOnceAsync() returns only a single utterance, it is suitable only for single shot recognition like command or query. For long-running multi-utterance recognition, use StartContinuousRecognitionAsync() instead.

pub async fn start_continuous_recognition_async<'_>(&'_ mut self) -> Result<()>

Asynchronously initiates continuous speech recognition operation.

pub async fn stop_continuous_recognition_async<'_>(&'_ mut self) -> Result<()>

Asynchronously terminates ongoing continuous speech recognition operation.

pub async fn start_keyword_recognition_async<'_>(
&'_ mut self,
model: KeywordRecognitionModel
) -> Result<()>

Asynchronously initiates keyword recognition operation.

pub async fn stop_keyword_recognition_async<'_>(&'_ mut self) -> Result<()>

Asynchronously terminates keyword recognition operation.

pub fn get_endpoint_id(&self) -> Result<String>

pub fn get_auth_token(&self) -> Result<String>

pub fn set_auth_token(&mut self, token: &str) -> Result<()>

Sets the authorization token that will be used for connecting to the service. Note: The caller needs to ensure that the authorization token is valid. Before the authorization token expires, the caller needs to refresh it by calling this setter with a new valid token. Otherwise, the recognizer will encounter errors during recognition.